requirements.txt
flask==3.0.0
python-telegram-bot==20.7
web3==6.11.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
